OLH-3072: Check for search terms before translating
We got loads of `safeTranslate` errors when we tried to release this
feature because the previous logic was always trying to translate the
additional search terms, even when we didn't have any. It was catching
those cases, but `req.t` has a side effect in logging an error if the
translation is missing.

Because we got lots of errors, this triggered our Cloudwatch alarms and
an automated canary rollback of the deploy.

This implements a hacky solution to check the translations to see if
this client has any additional serch terms first, then only doing the
translation if they do which will prevent the errors.

I've put the translations object at the base level of the file so it's
only created once on the first page load per container. This is the
'least bad' solution I could come up with on short notice (as we need to
release this feature). A better and (probably) more performant fix would
be to add a method to the registry package that exposes information
about a single client's translations. I'll write up a ticket to
investigate that option for later.