OLH-3025: Add an alarm for OIDC errors
Emit a new custom metric when we get an error on the OIDC callback, and
use that to trigger an alarm if we see too many.

I've looked at the production logs and right now we get 1 or 2
`access_denied` or `server_error` responses per day. I've configured the
alarm to trigger if 3 out of the last 5 minutes have more than 5 of
these errors. This should be well above the base threshold and so should
only trigger if there's a problem.