OLH-3057: Get users to enter their password before global logout
This journey is slightly different to the others in Home as we want to
show the global logout page first, then have the user enter their
password.

That means the series of HTTP requests that make up this journey now
look like:

```
GET  /security - click to "sign out of all apps"
GET  /global-logout - click the red button
POST /global-logout - redirect to password page
GET  /enter-password - user enters password
POST /enter-password - validates password, redirect to /global-logout/confirm
GET  /global-logout/confirm - sends the audit event, logs the user out
of Home
GET  / - redirects to sign in
... - user sees the sign in page

```

I chose to add a new route that contains the logout behaviour. That has
a couple of benefits:
1. It keeps the enter password controller behaviour the same as all
   other journeys
2. We can use the same state checks on the new route to make sure users
   have re-authenticated before they logout

The downside is that we have a destructive action triggered from a GET
request which isn't in the spirit of semantic HTTP. I think the benefits
outweigh this slight concern though.