OLH-3057: Use the performance test user for logout tests
We end all sessions for a user when we log them out. That meant that
when the tests are signed in to the default user and trigger a global
logout then it could sign out other tests which are doing other
scenarios. We think this is the cause of some test failures.