BAU: Use a UUID to trace requests
Previously we were using a combination of the journey ID (sessionId) and
the device ID (persistentSessionId) as the trace for all requests for a
user.

This worked fine for signed in users and allowed us to link requests to
service and other function calls.

However, is a user didn't have these IDs present in their session (ie.
if they aren't signed in) then we used a default value which makes it
impossible to distinguish between requests.

Additionally we couldn't follow a user from before to after signing in,
as their session IDs change, or aren't set before they sign in. We have
several pages like the contact page or the searchable list of services
which don't require the user to sign in, but we still need to be able to
follow traces to debug issues.

This switches the trace to a UUID which is saved in the user's session.
This will keep all functionality from before and also allow us to trace
requests before and after sign in, or if a user doesn't sign in at all.