BAU: Normalise the auth header when calling the MFA API
The Auth team have seen some errors on the MFA API related to the
format of the `Authorization` header we're sending.

Check the format is correct before we make the API request and throw an
error if it's not. This will make it easier for us to see when these
errors are happening.